Bulls_and_Cows
The old code-braking game.

An experiment on learning C# while collaborating by creating this old
game. The computer creates a randomly generated 4 digit array, and the
user has 12 attempts to try to figure the combination. If the user gets
1 correct number, he or she received a "bull" notification. If the user
selects a number that is within the array but it's improperly chosen, he
or she gets a "cow". All in all, the user will be uncertain of what he
or she got correct and wrong =-) Hopefuly a player can complete the game
within 12 turns and win.